requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
pymysql
python-dotenv
Werkzeug
pandas 
num2words
docx
python-docx
openpyxl
flask-wtf
flask-limiter
email_validator
itsdangerous 
flask-mail